# api-service/internal/items/http/app/docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  github_com_abdulazizax_mini-twitter_api-service_genproto_comment.Comment:
    properties:
      comment_serial:
        description: Serial number of the comment
        type: integer
      content:
        description: Comment content
        type: string
      created_at:
        allOf:
        - $ref: '#/definitions/timestamppb.Timestamp'
        description: Timestamp when the comment was created
      id:
        description: Comment ID
        type: string
      tweet_id:
        description: Tweet ID that this comment belongs to
        type: string
      updated_at:
        allOf:
        - $ref: '#/definitions/timestamppb.Timestamp'
        description: Timestamp when the comment was last updated
      username:
        description: Username of the comment's creator
        type: string
    type: object
  github_com_abdulazizax_mini-twitter_api-service_genproto_comment.GetCommentsForTweetResponse:
    properties:
      comments:
        description: List of comments for the tweet
        items:
          $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_comment.Comment'
        type: array
    type: object
...
  github_com_abdulazizax_mini-twitter_api-service_internal_models_tweet-handler.CreateCommentRequest:
    properties:
      content:
        type: string
      tweet_id:
        type: string
    type: object
...
  github_com_abdulazizax_mini-twitter_api-service_internal_models_tweet-handler.DeleteCommentRequest:
    properties:
      comment_id:
        type: string
      tweet_id:
        type: string
    type: object
...
  timestamppb.Timestamp:
    properties:
      nanos:
        description: |-
          Non-negative fractions of a second at nanosecond resolution. Negative
          second values with fractions must still have non-negative nanos values
          that count forward in time. Must be from 0 to 999,999,999
          inclusive.
        type: integer
      seconds:
        description: |-
          Represents seconds of UTC time since Unix epoch
          1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
          9999-12-31T23:59:59Z inclusive.
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Endpoints for MiniTwitter
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: '# MiniTwitter'
  version: 1.03.67.83.145
paths:
...
  /user/tweet/comment:
    post:
      consumes:
      - application/json
      description: Create a new comment for a tweet
      parameters:
      - description: Comment creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_internal_models_tweet-handler.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Create a new comment
      tags:
      - Comments
  /user/tweet/comment/{tweet_id}:
    get:
      description: Retrieve all comments for a specific tweet
      parameters:
      - description: Tweet ID
        in: path
        name: tweet_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_comment.GetCommentsForTweetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get comments for a tweet
      tags:
      - Comments
  /user/tweet/comment/delete:
    post:
      consumes:
      - application/json
      description: Delete a comment from a tweet
      parameters:
      - description: Comment deletion request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_internal_models_tweet-handler.DeleteCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - Comments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"